review comments and fix color of search field outline to match theme
